added support for sonarr and radarr custom scripts
support added for sonarr and radarr custom scripts. allows sonarr to set `-s true` and prbd will look for the `sonarr_episodefile_path` env variable. as well as radarr to set `-r true` and prbd will look for `radarr_moviefile_path` and prbd the file path provided. 
added `-log <path>` to log to a central location instead of just the folder the media file is located.